fix: subject y configTable
