+++ src/admin/hooks/hooksUI/index.ts
@@ -8,3 +8,7 @@ export { useListTeacherView } from "./Teacher/useListTeacherView";
 export { useCreateTeacherView } from "./Teacher/useCreateTeacherView";
 
 export { useListStudentView } from "./Student/useListStudentView";
+export { useCreateStudentView } from "./Student/useCreateStudentView";
+
+export { useListSubjectView } from "./Subject/useListSubjectView";
+export { useCreateSubjectView } from "./Subject/useCreateSubjectView";
