+++ src/admin/index.ts
@@ -1,2 +1,3 @@
 export * from "./hooks";
 export * from "./types";
+export * from "./tableConfig";
