+++ src/admin/tableConfig/index.ts
@@ -1,0 +1,5 @@
+export { getSubjectActions, getSubjectColumns } from "./subjectTable.config";
+export { getYearActions, getYearColumns } from "./yearTable.config";
+export { getTeacherActions, getTeacherColumns } from "./teacherTable.config";
+export { getStudentActions, getStudentColumns } from "./studentTable.config";
+export { getCourseActions, getCourseColumns } from "./courseTable.config";
