+++ src/admin/views/Subject/CreateSubjectView.tsx
@@ -1,123 +1,22 @@
-import { useEffect, useState } from "react";
-import { useParams, useNavigate } from "react-router-dom";
 import { motion } from "framer-motion";
 import { FaGraduationCap, FaSave } from "react-icons/fa";
-import BooleanInput from "../../../shared/components/BooleanInput/BooleanInputComponent";
-import Button from "../../../shared/components/Button/ButtonComponent";
-import Card from "../../../shared/components/Card/CardComponent";
-import Input from "../../../shared/components/Input/InputComponent";
-import { useYear } from "../../hooks/useYear";
-import { useCourse } from "../../hooks/useCourse";
-import { useTeacher } from "../../hooks/useTeacher";
-import { useSubject } from "../../hooks/useSubject";
-import { useToaster } from "../../../shared/hooks/useToaster";
+import { BooleanInput, Button, Card, Input } from "@/shared";
 import styles from "./CreateSubjectView.module.css";
+import { useCreateSubjectView } from "@/admin";
 
 const CreateSubjectView: React.FC = () => {
-  const { years, getYear } = useYear();
-  const { courses, getCourse } = useCourse();
-  const { teacher, getTeacher } = useTeacher();
-  const { loading, selectedSubject, registerSubject, updateSubject } =
-    useSubject();
-  const { id } = useParams<{ id: string }>();
-  const { showToast } = useToaster();
-  const navigate = useNavigate();
-
-  const [formData, setFormData] = useState({
-    name: "",
-    yearId: 0,
-    courseId: 0,
-    teacherId: 0,
-    optional: false,
-  });
-
-  const isEditMode = Boolean(id);
-
-  const filteredCourses = courses.filter(
-    (course) => course.year.id === formData.yearId
-  );
-
-  const resetFormData = () => {
-    setFormData({
-      name: "",
-      yearId: 0,
-      courseId: 0,
-      teacherId: 0,
-      optional: false,
-    });
-  };
-
-  useEffect(() => {
-    const loadInitialData = async () => {
-      await Promise.all([getYear(), getCourse(), getTeacher()]);
-    };
-
-    loadInitialData();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  useEffect(() => {
-    if (isEditMode && selectedSubject) {
-      setFormData({
-        name: selectedSubject.name || "",
-        yearId: selectedSubject.course?.year?.id || 0,
-        courseId: selectedSubject.course?.id || 0,
-        teacherId: selectedSubject.teacher?.id || 0,
-        optional: selectedSubject.optional || false,
-      });
-    }
-  }, [isEditMode, selectedSubject]);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    const payload = {
-      name: formData.name.trim(),
-      courseId: formData.courseId,
-      teacherId: formData.teacherId === 0 ? null : formData.teacherId,
-      optional: formData.optional,
-    };
-    try {
-      if (isEditMode && id) {
-        await updateSubject({
-          id: Number(id),
-          ...payload,
-        });
-        showToast({
-          title: "Materia actualizada exitosamente",
-          message: "La materia ha sido actualizada exitosamente",
-          type: "success",
-          position: "bottom-right",
-        });
-        navigate("/dashboard/subjects/list");
-      } else {
-        await registerSubject(payload);
-        showToast({
-          title: "Materia creada exitosamente",
-          message: "La materia ha sido creada exitosamente",
-          type: "success",
-          position: "bottom-right",
-        });
-        resetFormData();
-      }
-    } catch (err) {
-      showToast({
-        title:
-          "Error al" + (isEditMode ? " actualizar" : " crear") + " la materia",
-        type: "error",
-        position: "bottom-right",
-      });
-    }
-  };
-
-  const handleChange = (field: string, value: string | number | boolean) => {
-    setFormData((prev) => ({ ...prev, [field]: value }));
-  };
-
-  const handleCancel = () => {
-    if (isEditMode) {
-      navigate("/dashboard/subjects/list");
-    }
-  };
+  const {
+    loading,
+    years,
+    teacher,
+    isEditMode,
+    formData,
+    filteredCourses,
+    setFormData,
+    handleChange,
+    handleSubmit,
+    handleCancel,
+  } = useCreateSubjectView();
 
   return (
     <motion.div
