+++ src/admin/views/Subject/ListSubjectView.tsx
@@ -1,186 +1,38 @@
-import type React from "react";
-import { useEffect } from "react";
-import { useNavigate } from "react-router-dom";
 import { motion } from "framer-motion";
-import { FaCalendarAlt, FaEdit, FaTrash, FaPlus, FaUser } from "react-icons/fa";
-import type { SubjectResponseDto } from "../../services/subject/SubjectService";
-import Button from "../../../shared/components/Button/ButtonComponent";
-import { DataTable } from "../../../shared/components/DataTable";
-import type {
-  DataTableColumn,
-  DataTableAction,
-} from "../../../shared/components/DataTable";
-import { useSubject } from "../../hooks/useSubject";
-import { useConfirmation } from "../../../shared/hooks/useConfirmation";
-import { useToaster } from "../../../shared/hooks/useToaster";
-import usePaginationParams from "../../../shared/hooks/usePaginateParams";
+import { FaCalendarAlt, FaPlus } from "react-icons/fa";
+import {
+  getSubjectActions,
+  getSubjectColumns,
+  useListSubjectView,
+} from "@/admin";
+import { Button, DataTable, itemVariants, containerVariants } from "@/shared";
 import styles from "./ListSubjectView.module.css";
 
 const ListSubjectView: React.FC = () => {
   const {
-    loading,
+    // states
     paginatedSubjects,
-    setSelectedSubject,
-    getPaginatedSubject,
-    deleteSubject,
-  } = useSubject();
-  const {
+    loading,
+
+    // pagination
     paginationParams,
     handleSearch,
     handleSort,
     handlePageChange,
     handlePageSizeChange,
-  } = usePaginationParams();
-  const { showConfirmation } = useConfirmation();
-  const { showToast } = useToaster();
-  const navigate = useNavigate();
 
-  useEffect(() => {
-    const loadPaginatedSubjects = async () => {
-      await getPaginatedSubject(paginationParams);
-    };
-    loadPaginatedSubjects();
-  }, [paginationParams, getPaginatedSubject]);
+    // actions
+    handleEdit,
+    handleDelete,
+    navigate,
+  } = useListSubjectView();
 
-  const handleEdit = (subject: SubjectResponseDto) => {
-    showConfirmation({
-      title: "Modificar Materia",
-      message: `¿Está seguro que desea modificar la materia "${subject.name}"?`,
-      type: "warning",
-      onConfirm: () => {
-        setSelectedSubject(subject);
-        navigate(`/dashboard/subjects/edit/${subject.id}`);
-      },
-    });
-  };
-
-  const handleDelete = (subject: SubjectResponseDto) => {
-    showConfirmation({
-      title: "Eliminar Materia",
-      message: `¿Está seguro que desea eliminar la materia "${subject.name}"?`,
-      type: "danger",
-      showDoubleConfirmation: true,
-      onConfirm: async () => {
-        await deleteSubject(subject.id);
-        await getPaginatedSubject(paginationParams);
-        showToast({
-          title: "Materia eliminada exitosamente",
-          message: "La materia ha sido eliminada exitosamente",
-          type: "success",
-          position: "bottom-right",
-        });
-      },
-    });
-  };
-
-  // Definición de columnas para la tabla
-  const columns: DataTableColumn<SubjectResponseDto>[] = [
-    {
-      key: "id",
-      label: "ID",
-      sortable: true,
-      width: "100px",
-      className: styles.idColumn,
-      render: (subject) => <span className={styles.idBadge}>{subject.id}</span>,
-    },
-    {
-      key: "name",
-      label: "Nombre",
-      sortable: true,
-      className: styles.nameColumn,
-      render: (subject) => (
-        <div className={styles.nameCell}>
-          <div className={styles.nameWrapper}>
-            <span>{subject.name}</span>
-          </div>
-        </div>
-      ),
-    },
-    {
-      key: "year",
-      label: "Año",
-      sortable: true,
-      className: styles.nameColumn,
-      render: (subject) => (
-        <div className={styles.nameCell}>
-          <div className={styles.nameWrapper}>
-            <span>{subject.course?.year?.name || "Sin asignar"}</span>
-          </div>
-        </div>
-      ),
-    },
-    {
-      key: "course",
-      label: "Curso",
-      sortable: true,
-      className: styles.nameColumn,
-      render: (subject) => (
-        <div className={styles.nameCell}>
-          <div className={styles.nameWrapper}>
-            <span>{subject.course?.name || "Sin asignar"}</span>
-          </div>
-        </div>
-      ),
-    },
-    {
-      key: "teacher",
-      label: "Docente",
-      sortable: true,
-      className: styles.nameColumn,
-      render: (subject) => (
-        <div className={styles.nameCell}>
-          <div className={styles.nameWrapper}>
-            <FaUser
-              className={
-                subject.teacher?.name
-                  ? styles.teacherIcon
-                  : styles.teacherIconUnassigned
-              }
-            />
-            <span>
-              {subject.teacher?.name + " " + subject.teacher?.lastname ||
-                "Sin asignar"}
-            </span>
-          </div>
-        </div>
-      ),
-    },
-  ];
-
-  // Definición de acciones para la tabla
-  const actions: DataTableAction<SubjectResponseDto>[] = [
-    {
-      label: "Editar",
-      icon: <FaEdit />,
-      onClick: handleEdit,
-      variant: "ghost",
-      className: styles.editButton,
-      title: "Modificar materia",
-    },
-    {
-      label: "Eliminar",
-      icon: <FaTrash />,
-      onClick: handleDelete,
-      variant: "ghost",
-      className: styles.deleteButton,
-      title: "Eliminar materia",
-    },
-  ];
-
-  const containerVariants = {
-    hidden: { opacity: 0 },
-    visible: {
-      opacity: 1,
-      transition: {
-        staggerChildren: 0.1,
-      },
-    },
-  };
-
-  const itemVariants = {
-    hidden: { y: 20, opacity: 0 },
-    visible: { y: 0, opacity: 1 },
-  };
+  const columns = getSubjectColumns(styles);
+  const actions = getSubjectActions({
+    styles,
+    handleEdit,
+    handleDelete,
+  });
 
   return (
     <motion.div
