+++ src/shared/components/index.ts
@@ -3,3 +3,4 @@ export { default as DataTable } from "./DataTable/DataTableComponent";
 export type { DataTableColumn, DataTableAction } from "./DataTable";
 export { default as Card } from "./Card/CardComponent";
 export { default as Input } from "./Input/InputComponent";
+export { default as BooleanInput } from "./BooleanInput/BooleanInputComponent";
